Api added

+++ Data/Services/AppleIdService.cs
@@ -1,3 +1,4 @@
+using AppleAccounts.Controllers;
 using AppleAccounts.Data.Enums;
 using AppleAccounts.Models;
 using Microsoft.EntityFrameworkCore;
@@ -19,18 +20,6 @@ namespace AppleAccounts.Data.Services
             return result;
         }
 
-        public async Task<IEnumerable<AppleId>> GetAppleIds(int status, bool expired)
-
-        {
-            return await _context.AppleIds.Where(i => i.Status == (AppleIdStatus)status && i.Expired == expired).ToListAsync();
-        }
-
-        public async Task<IEnumerable<AppleId>> GetAppleIds(bool expired)
-
-        {
-            return await _context.AppleIds.Where(i => i.Expired == expired).ToListAsync();
-        }
-
         public async Task<AppleId> UpdateAsync(int id, AppleId newAppleId)
         {
             _context.Update(newAppleId);
@@ -41,15 +30,25 @@ namespace AppleAccounts.Data.Services
         public async Task DeleteAsync(int id)
         {
             var result = await _context.AppleIds.FirstOrDefaultAsync(n => n.Id == id);
-            _context.AppleIds.Remove(result);
-            await _context.SaveChangesAsync();
+            if (result != null)
+            {
+                _context.AppleIds.Remove(result);
+                await _context.SaveChangesAsync();
+            }
         }
 
-        public async Task<IEnumerable<AppleId>> GetAppleIds(string searchPhrase)
+        public async Task<IEnumerable<AppleId>> GetAppleIds(bool expired, string? status = null, string? filterQuery = "")
         {
-            return await _context.AppleIds.Where(i =>
-                i.Email.Contains(searchPhrase) || i.Id.ToString().Contains(searchPhrase))
-                .ToListAsync();
+            IQueryable<AppleId> appleIds = _context.AppleIds.AsQueryable().Where(i => i.Expired == expired);
+            if (!string.IsNullOrEmpty(filterQuery))
+                appleIds = appleIds.Where(i => i.Email.Contains(filterQuery) || i.Id.ToString().Contains(filterQuery));
+
+            if (!string.IsNullOrWhiteSpace(status))
+            {
+                if (int.TryParse(status, out int id))
+                    appleIds = appleIds.Where(i => i.Status == (AppleIdStatus)id);
+            }
+            return await appleIds.ToListAsync();
         }
     }
 }
